[FFM-1928]: Increment the version of the Java FF SDK (#29311)
This changes the version of the FF SDK to 1.0.5.  We have observed issues
with dependenacies on our latest SDK, but 1.0.5 contains fixes to address
deadlock issues that can occur.

We have also added a waitForIinit block that will wait for the SDK to complete
intialization before reutning and clearly log if we do not.

(cherry picked from commit 691fab4508a9b960aaf07316b2645ce84f3d85aa)