[fix]: CI-3384: [TI] addon should switch between -DargLine and -DharnessArgLine
This change introduces a helper method to check the presence of harnessArgLine tag in any XML
file. If such a tag exists, the -DArgLine will be replaced by -DharnessArgLine. This is needed
to support attaching TI agent as well as a customer chosen agent such as JaCoCo for code
coverage.